refactor(sdk-keys): drop rtk_server_side_sdk_keys flag wrapper
Inline the RTK-based SDK keys page into web/components/pages/sdk-keys/SDKKeysPage.tsx
and remove the SDKEnvironmentKeysSettings indirection. The flag has shipped, so the
wrapper that toggled between legacy and RTK paths is no longer needed.

Refs #7140
